+++ e3nn/util/_argtools.py
@@ -97,14 +97,41 @@ def _get_device(mod: torch.nn.Module) -> torch.device:
     return a_buf.device if a_buf is not None else 'cpu'
 
 
-def _to_device(args, device):
+def _get_floating_dtype(mod: torch.nn.Module) -> torch.dtype:
+    """Guess floating dtype for module.
+
+    Assumes no mixed precision.
+    """
+    FLOATING_DTYPES = [torch.float32, torch.float64]
+    # Try to a get a parameter
+    a_buf = None
+    for buf in mod.parameters():
+        if buf.dtype in FLOATING_DTYPES:
+            a_buf = buf
+            break
+    if a_buf is None:
+        # If there isn't one, try to get a buffer
+        for buf in mod.buffers():
+            if buf.dtype in FLOATING_DTYPES:
+                a_buf = buf
+                break
+    return a_buf.dtype if a_buf is not None else torch.get_default_dtype()
+
+
+def _to_device_dtype(args, device=None, dtype=None):
+    kwargs = {}
+    if device is not None:
+        kwargs['device'] = device
+    if dtype is not None:
+        kwargs['dtype'] = dtype
+
     if isinstance(args, torch.Tensor):
-        return args.to(device=device)
+        return args.to(**kwargs)
     elif isinstance(args, tuple):
-        return tuple(_to_device(e, device) for e in args)
+        return tuple(_to_device_dtype(e, **kwargs) for e in args)
     elif isinstance(args, list):
-        return [_to_device(e, device) for e in args]
+        return [_to_device_dtype(e, **kwargs) for e in args]
     elif isinstance(args, dict):
-        return{k: _to_device(v, device) for k, v in args.items()}
+        return{k: _to_device_dtype(v, **kwargs) for k, v in args.items()}
     else:
-        raise TypeError("Only (nested) dict/tuple/lists of Tensors can be moved to a device.")
+        raise TypeError("Only (nested) dict/tuple/lists of Tensors can be moved to a device/dtype.")