+++ e3nn/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.2.7"
+__version__ = "0.2.8"
 
 
 from typing import Dict