+++ ChangeLog.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Most recent change on the bottom.
 
 ## [Unreleased]
+
+## [0.2.8] - 2021-04-21
 ### Added
 - `squared` option to `o3.Norm`
 - `e3nn.nn.models.v2104.voxel_convolution.Convolution` made to be resolution agnostic